pyproject.toml
[tool.black]
line-length=79
target-version=["py38"]

[tool.isort]
atomic=true
line_length = 79
lines_after_imports=2
lines_between_types=1
use_parentheses = true
balanced_wrapping = true
include_trailing_comma = true
multi_line_output = 3
known_third_party=["pytest", "fastapi", "httpx", "yarl"]
known_first_party=["unsupported_pcidevs"]

[tool.coverage.run]
branch = true
source = ["unsupported_pcidevs"]

[tool.coverage.paths]
source = ["unsupported_pcidevs/"]

[tool.coverage.report]
fail_under = 100
skip_covered = true
show_missing = true

[tool.pytest.ini_options]
minversion = "6.0.0"
#addopts = "-vv -s --cov --no-cov-on-fail --tb=native"
addopts = "-vv -s --no-cov-on-fail --tb=native"
testpaths = "tests/"
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "| %(asctime)s | %(name)s | %(levelname)s | %(filename)s | %(message)s"

[tool.interrogate]
ignore-init-method = true
ignore-init-module = true
ignore-magic = false
ignore-semiprivate = false
ignore-private = false
ignore-property-decorators = false
ignore-module = true
fail-under = 0
exclude = ["setup.py", "docs", "build", "tests/load_testing"]
ignore-regex = ["^get$", "^mock_.*", ".*BaseClass.*", "^get.*$", "^factory$"]
verbose = 2
quiet = false
whitelist-regex = []
color = true

[tool.poetry]
name = "cv"
version = "0.1.0"
description = ""
authors = ["ZhukovGreen <zhukovgreen@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.6"
pygithub = "^1.46"
envparse = "^0.2.0"
gitpython = "^3.0.8"
mkdocs = "^1.0.4"
markdown-include = "^0.5.1"
mkdocs-material = "^4.6.3"
weasyprint = "^51"

[tool.poetry.dev-dependencies]
pytest = "^6.0.0"
neovim = "^0.3.1"

[build-system]
requires = ["poetry>=1.0.3"]
build-backend = "poetry.masonry.api"
